build QGIS Docker on cron jobs

+++ .ci/travis/linux/script.sh
@@ -15,8 +15,21 @@
 
 set -e
 
-DIR=$(git rev-parse --show-toplevel)/.docker
+
 
 mkdir -p $CCACHE_DIR
 
-docker-compose -f $DOCKER_COMPOSE run --rm qgis-deps
+if [[ $DOCKER_QGIS_IMAGE_PUSH =~ true ]]; then
+  DIR=$(git rev-parse --show-toplevel)/.docker
+  pushd ${DIR}
+  echo "${bold}Building QGIS Docker image...${endbold}"
+  docker build -v /root/QGIS:/root/QGIS --cache-from "qgis/qgis:${DOCKER_TAG}" -t "qgis/qgis:${DOCKER_TAG}" -f qgis.dockerfile .
+  if [[ $DOCKER_QGIS_IMAGE_PUSH =~ true ]]; then
+    echo "${bold}Pushing image to docker hub...${endbold}"
+    docker login -u="$DOCKER_USERNAME" -p="$DOCKER_PASSWORD"
+    docker push "qgis/qgis:${DOCKER_TAG}"
+  fi
+  popd
+else
+  docker-compose -f $DOCKER_COMPOSE run --rm qgis-deps
+fi
